+++ src/main/webapp/assets/css/myhome/myhome-delete-ok.css
@@ -21,163 +21,13 @@
         flex-direction: column;
     }
 
-    /* 헤더 부분 */
-
-    .header-container {
-        display: flex;
-        flex-direction: column;
-        width: 100%;
-    }
-
-    #header-wrap {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        width: 1440px;
-        height: 151px;
-        margin: 0 auto;
-    }
-
-    #header-wrap #border {
-        display: flex;
-        flex-direction: column;
-        margin: 0 auto;
-    }
-
-    #header-wrap #nav {
-        position: relative;
-        display: flex;
-        width: 1440px;
-        margin: 0 auto;
-        /* 헤더 윗공간 조절 */
-        margin-top: 20px;
-        justify-content: space-between;
-        align-items: end;
-        margin-bottom: 15px;
-    }
-
-    #header-wrap #nav .sector {
-        position: absolute;
-        line-height: 20px;
-        margin-left: 10px;
-        /* border: 1px solid #818181; */
-        background-color: #818181;
-        height: 55px;
-        width: 0.1px;
-        top: 34px;
-        right: 96px;
-    }
-
-    #header-wrap #nav .left {
-        width: 236px;
-        height: 60px;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-    }
-
-    #header-wrap #nav .left img {
-        width: 236px;
-        height: 60px;
-    }
-
-    #header-wrap #nav .right {
-        display: flex;
-        gap: 26px;
-        align-items: end;
-        padding-top: 44px;
-    }
-
-    #header-wrap #nav .right .icons {
-        display: flex;
-        gap: 26px;
-    }
-
-    #header-wrap #nav .right .icons img {
-        display: flex;
-        width: 30px;
-        height: 30px;
-    }
-
-    #header-wrap #nav .right .login {
-        display: flex;
-        font-size: 16px;
-        margin-left: 26px;
-    }
-
-    #header-wrap #nav .right .login a {
-        display: flex;
-        width: 100%;
-        height: 100%;
-        text-align: center;
-        justify-content: center;
-        text-decoration-line: none;
-        color: black;
-        padding-right: 31px;
-    }
-
-    /* 메뉴부분 */
-
-    .menu-container {
-        display: flex;
-        flex-direction: column;
-        flex: 1;
-        z-index: 10;
-    }
-
-    #sector {
-        display: flex;
-        width: 100%;
-        height: 0.1px;
-        background-color: #818181;
-        z-index: 50;
-    }
-
-    .menu-wrap {
-        display: flex;
-        height: 50px;
-        font-weight: bold;
-        border-top: solid #818181 0.1px;
-        border-bottom: solid #818181 0.1px;
-    }
-
-    .menu-wrap .menu {
-        display: flex;
-        align-items: center;
-        width: 150px;
-        height: 50px;
-    }
-
-    .menu-wrap .menu a:hover {
-        color: #FFC303;
-    }
-
-    .sub-menu-wrap {
-        display: flex;
-    }
-
-    .sub-menu-wrap ul {
-        overflow: hidden;
-        height: 0;
-    }
-
-    .sub-menu-wrap .sub-menu {
-        width: 150px;
-        height: 50px;
-        display: flex;
-        align-items: center;
-        text-align: center;
-    }
-
-    .sub-menu-wrap .sub-menu a:hover {
-        color: #FFC303;
-    }
-
+   
     /* contents */
 
     #unsubscribe-container {
         display: block;
         margin-top: 100px;
+        margin-bottom: 187px;
     }
 
     .unsubscribe-text {
